^ Refactored chunk of code
